% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinityPropagation.R
\name{affinityPropagation}
\alias{affinityPropagation}
\title{Affinity Propagation}
\usage{
affinityPropagation(idsInSet, score)
}
\arguments{
\item{idsInSet}{A list of set names and their member IDs.}

\item{score}{A vector of addible scores with the same length used to assign input preference;
higher score has larger weight, i.e. -logP.}
}
\value{
A list of \code{clusters} and \code{representatives} for each cluster.
\describe{
 \item{clusters}{A list of character vectors of set IDs in each cluster.}
 \item{representatives}{A character vector of representatives for each cluster.}
}
}
\description{
Use affinity propagation to cluster similar gene sets to reduce redundancy in report.
}
\author{
Zhiao Shi, Yuxing Liao
}
