% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\name{cacheUrl}
\alias{cacheUrl}
\title{cacheUrl}
\usage{
cacheUrl(dataUrl, cache = NULL, query = NULL)
}
\arguments{
\item{dataUrl}{The URL of data}

\item{cache}{The cache directory. Defaults to \code{NULL} and fetches data from server.}

\item{query}{The list of queries passed on to httr methods}
}
\value{
response object from httr request
}
\description{
Get data from a URL or cache and optionally save in cache for reuse
}
\keyword{internal}
