% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNtaReport.R
\name{createNtaReport}
\alias{createNtaReport}
\title{Create HTML Report for NTA}
\usage{
createNtaReport(networkName, method, sigMethod, fdrThr, topThr,
  highlightType, outputDirectory, projectDir, projectName, hostName)
}
\description{
Create HTML Report for NTA
}
\keyword{internal}
