% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichResultSection.R
\name{enrichResultSection}
\alias{enrichResultSection}
\title{enrichResultSection}
\usage{
enrichResultSection(
  enrichMethod,
  enrichedSig,
  geneSet,
  geneSetDes,
  geneSetDag,
  geneSetNet,
  clusters
)
}
\description{
Conditionally render template of main result section. Actual work is carried out in front end
}
\keyword{internal}
