% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagUtils.R
\name{expandDag}
\alias{expandDag}
\title{Expand enriched GO IDs to include ancestors up to the root}
\usage{
expandDag(goTermList, dagEdgeList)
}
\description{
Returns expanded nodes and DAG tree edges
}
\keyword{internal}
