% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fill_input_data_frame}
\alias{fill_input_data_frame}
\title{Fill relation data frame for GSEA input}
\usage{
fill_input_data_frame(gmt, genes, gene_sets)
}
\arguments{
\item{gmt}{A Data Frame with geneSet and gene columns from the GMT file}

\item{genes}{A vector of genes}

\item{gene_sets}{A vector of gene sets}
}
\value{
A Data Frame with the first column of gene and 1 or 0 for other columns of gene sets.
}
\description{
Fill 1 for gene in gene set
}
\details{
See https://github.com/extendr/extendr/issues/612 for how to export DataFrame

## Diagram
```shell
        Gene Sets
      ┌───────────┐  First column named 'gene' containing gene name
      │A0100110100│  1 = in set
Genes │B0100101000│  0 = not in set
      │C1011101001│  Due to limitiations with extendr-api v 0.6.0,
      └───────────┘  function returns a list, and the R package will
                     add the first 'gene' column
```
}
\author{
John Elizarraras
}
\keyword{internal}
