% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fillInputDataFrame}
\alias{fillInputDataFrame}
\title{Fill relation data frame for GSEA input}
\usage{
fillInputDataFrame(gmt, genes, geneSets)
}
\arguments{
\item{gmt}{A Data Frame with geneSet and gene columns from the GMT file}

\item{genes}{A vector of genes}

\item{geneSets}{A vector of gene sets}
}
\value{
A Data Frame with the first column of gene and 1 or 0 for other columns of gene sets.
}
\description{
Fill 1 for gene in gene set
}
\author{
Yuxing Liao
}
\keyword{internal}
