% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{gsea_rust}
\alias{gsea_rust}
\title{Run GSEA using rust library}
\usage{
gsea_rust(
  min_overlap,
  max_overlap,
  permutations,
  sets,
  parts,
  analytes,
  ranks,
  seed
)
}
\arguments{
\item{min_overlap}{the minimum overlap between analyte set and analyte list}

\item{max_overlap}{the maximum overlap between analyte set and analyte list}

\item{permutations}{the number of permutations to run}

\item{sets}{A vector of analyte set names}

\item{parts}{A list of the analytse in the analyte sets}

\item{analytes}{A vector of analytes names in the GSEA list}

\item{ranks}{A vector of ranks for the analytes in the GSEA list}

\item{seed}{Seed for the permutation RNG, or \code{NA} to draw from system entropy}
}
\value{
List of the results of GSEA
}
\description{
Run GSEA using rust library
}
\author{
John Elizarraras
}
\keyword{internal}
