% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gseaPermutation}
\alias{gseaPermutation}
\title{Permutaion in GSEA algorithm}
\usage{
gseaPermutation(inset_scores, outset_scores, expression_value)
}
\arguments{
\item{inset_scores}{Scaled score matrix for genes in sets}

\item{outset_scores}{Normalized score matrix for genes not in sets}

\item{expression_value}{Vector of gene rank scores}
}
\value{
A vector of concatenated random minimal,maimum and best running sum scores for each set.
}
\description{
Permutaion in GSEA algorithm
}
\author{
Yuxing Liao
}
\keyword{internal}
