% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kMedoid.R
\name{kMedoid}
\alias{kMedoid}
\title{kMedoid}
\usage{
kMedoid(idsInSet, score, maxK = 10)
}
\arguments{
\item{idsInSet}{a list of sets of ids}

\item{score}{a vector of scores for each set}

\item{maxK}{maximum number of clusters}
}
\description{
kMedoid clustering
}
