% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkModification.R
\name{linkModification}
\alias{linkModification}
\title{Modify the link to highlight the genes in the pathways}
\usage{
linkModification(enrichMethod, enrichDatabase, enrichPathwayLink, geneList,
  interestingGeneMap)
}
\description{
Currently, we only have wikipathway and kegg pathways that need to modify the link
}
\keyword{internal}
