\name{listArchiveURL}
\alias{listArchiveURL}

\title{
	List the server URL of each version of the data update 
}

\description{
	The listArchiveURL function can list the server URL of 
	each version of the data update. Users can select 
	different server URL as the \code{hostName} to perform 
	the enrichment analysis.
}

\usage{
listArchiveURL()
}

\author{
	Jing Wang
}

\value{
	The listArchiveURL function can list the server URL of 
	each version of the data update.
}

\examples{
	#archiveURL <- listArchiveURL()
}

\keyword{ methods }
