% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaEnrichResultSection.R
\name{metaEnrichResultSection}
\alias{metaEnrichResultSection}
\title{metaEnrichResultSection}
\usage{
metaEnrichResultSection(
  enrichMethod,
  enrichedSig,
  geneSet,
  geneSetDes,
  geneSetDag,
  geneSetNet,
  clusters
)
}
\description{
Conditionally render template of main result section. Actual work is carried out in front end
}
\keyword{internal}
