% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaLinkModification.R
\name{metaLinkModification}
\alias{metaLinkModification}
\title{Modify the link to highlight the genes in the pathways}
\usage{
metaLinkModification(
  enrichMethod,
  enrichPathwayLink,
  geneList_list,
  interestingGeneMap_list,
  hostName = "https://www.webgestalt.org/",
  geneSet
)
}
\description{
Currently, we only have wikipathway and kegg pathways that need to modify the link
}
\keyword{internal}
