% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{nta_rust}
\alias{nta_rust}
\title{Calculate random walk permutations for a network from seeds}
\usage{
nta_rust(edge_list, seeds)
}
\arguments{
\item{edge_list}{A list of edges}

\item{seeds}{A list of seeds}
}
\value{
A list of nodes and scores
}
\description{
Calculate random walk permutations for a network from seeds
}
\author{
John Elizarraras
}
\keyword{internal}
