% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_multiomics_ora}
\alias{rust_multiomics_ora}
\title{Run multiomics ORA using Rust library}
\usage{
rust_multiomics_ora(sets, big_part_vec, interest, reference, method)
}
\arguments{
\item{sets}{list of  the names of the analyte sets}

\item{big_part_vec}{list of the analyte in the analyte sets}

\item{interest}{list of analytes of interest}

\item{reference}{list of analytes in the reference set}

\item{method}{meta-analysis method to get meta-p values}
}
\value{
A list of vectors containing the results of ORA, with each list corresponding to each input list
}
\description{
Run multiomics ORA using Rust library
}
\author{
John Elizarraras
}
\keyword{internal}
