% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryDescription.R
\name{specificParameterSummaryGsea}
\alias{specificParameterSummaryGsea}
\title{specificParameterSummaryGsea}
\usage{
specificParameterSummaryGsea(organism, interestingGeneMap, geneSet, minNum,
  maxNum, sigMethod, fdrThr, topThr, perNum, enrichedSig, reportNum)
}
\description{
Render job summary section of GSEA specific parameters
}
\keyword{internal}
