% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryDescription.R
\name{specificParameterSummaryOra}
\alias{specificParameterSummaryOra}
\title{specificParameterSummaryOra}
\usage{
specificParameterSummaryOra(
  organism,
  referenceGeneList,
  geneSet,
  referenceGeneFile,
  referenceGene,
  referenceGeneType,
  referenceSet,
  minNum,
  maxNum,
  sigMethod,
  fdrThr,
  topThr,
  fdrMethod,
  enrichedSig,
  reportNum,
  repAdded,
  numAnnoRefUserId,
  interestingGeneMap,
  hostName
)
}
\description{
Render job summary section of ORA specific parameters
}
\keyword{internal}
