% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryDescription.R
\name{summaryDescription}
\alias{summaryDescription}
\title{summaryDescription}
\usage{
summaryDescription(projectName, organism, interestGeneFile, interestGene,
  interestGeneType, enrichMethod, enrichDatabase, enrichDatabaseFile,
  enrichDatabaseType, enrichDatabaseDescriptionFile, interestingGeneMap,
  referenceGeneList, referenceGeneFile, referenceGene, referenceGeneType,
  referenceSet, minNum, maxNum, sigMethod, fdrThr, topThr, fdrMethod,
  enrichedSig, reportNum, perNum, geneSet, hostName)
}
\description{
Render job summary section
}
\keyword{internal}
