% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebGestaltRMultiOmics.R
\name{WebGestaltRMultiOmics}
\alias{WebGestaltRMultiOmics}
\title{WebGestaltRMultiOmics}
\usage{
WebGestaltRMultiOmics(
  analyteLists = NULL,
  analyteListFiles = NULL,
  analyteTypes = NULL,
  enrichMethod = "ORA",
  organism = "hsapiens",
  enrichDatabase = NULL,
  enrichDatabaseFile = NULL,
  enrichDatabaseType = NULL,
  enrichDatabaseDescriptionFile = NULL,
  collapseMethod = "mean",
  minNum = 10,
  maxNum = 500,
  fdrMethod = "BH",
  sigMethod = "fdr",
  fdrThr = 0.05,
  topThr = 10,
  reportNum = 100,
  setCoverNum = 10,
  perNum = 1000,
  gseaP = 1,
  isOutput = TRUE,
  outputDirectory = getwd(),
  projectName = NULL,
  dagColor = "binary",
  saveRawGseaResult = FALSE,
  gseaPlotFormat = "png",
  nThreads = 1,
  cache = NULL,
  hostName = "https://www.webgestalt.org/",
  useWeightedSetCover = TRUE,
  useAffinityPropagation = FALSE,
  usekMedoid = FALSE,
  kMedoid_k = 25,
  isMetaAnalysis = TRUE,
  mergeMethod = "mean",
  normalizationMethod = "rank",
  referenceLists = NULL,
  referenceListFiles = NULL,
  referenceTypes = NULL,
  referenceSets = NULL,
  listNames = NULL,
  seed = NULL
)
}
\arguments{
\item{analyteLists}{\code{vector} of the ID type of the corresponding interesting analyte list. The supported ID types of
WebGestaltR for the selected organism can be found by the function \code{listIdType}. If
the \code{organism} is \code{others}, users do not need to set this parameter. The length of \code{analyteLists} should be
the same as the length of \code{analyteListFiles} or \code{analyteLists}.}

\item{analyteListFiles}{If \code{enrichMethod} is \code{ORA}, the extension of
the \code{analyteListFiles} should be \code{txt} and each file can only contain one column:
the interesting analyte list. If \code{enrichMethod} is \code{GSEA}, the extension of the
\code{analyteListFiles} should be \code{rnk} and the files should contain two columns
separated by tab: the analyte list and the corresponding scores.}

\item{analyteTypes}{a vector containing the ID types of the analyte lists.}

\item{enrichMethod}{Enrichment methods: \code{ORA}or \code{GSEA}.}

\item{organism}{Currently, WebGestaltR supports 12 organisms. Users can use the function
\code{listOrganism} to check available organisms. Users can also input \code{others} to
perform the enrichment analysis for other organisms not supported by WebGestaltR. For
other organisms, users need to provide the functional categories, interesting list and
reference list (for ORA method). Because WebGestaltR does not perform the ID mapping for
the other organisms, the above data should have the same ID type.}

\item{enrichDatabase}{The functional categories for the enrichment analysis. Users can use
the function \code{listGeneSet} to check the available functional databases for the
selected organism. Multiple databases in a vector are supported for ORA and GSEA.}

\item{enrichDatabaseFile}{Users can provide one or more GMT files as the functional
category for enrichment analysis. The extension of the file should be \code{gmt} and the
first column of the file is the category ID, the second one is the external link for the
category. Genes annotated to the category are from the third column. All columns are
separated by tabs. The GMT files will be combined with \code{enrichDatabase}.}

\item{enrichDatabaseType}{The ID type of the genes in the \code{enrichDatabaseFile}.
If users set \code{organism} as \code{others}, users do not need to set this ID type because
WebGestaltR will not perform ID mapping for other organisms. The supported ID types of
WebGestaltR for the selected organism can be found by the function \code{listIdType}.}

\item{enrichDatabaseDescriptionFile}{Users can also provide description files for the custom
\code{enrichDatabaseFile}. The extension of the description file should be \code{des}. The
description file contains two columns: the first column is the category ID that should be
exactly the same as the category ID in the custom \code{enrichDatabaseFile} and the second
column is the description of the category. All columns are separated by tabs.}

\item{collapseMethod}{The method to collapse duplicate IDs with scores. \code{mean},
\code{median}, \code{min} and \code{max} represent the mean, median, minimum and maximum
of scores for the duplicate IDs.}

\item{minNum}{WebGestaltR will exclude the categories with the number of annotated genes
less than \code{minNum} for enrichment analysis. The default is \code{10}.}

\item{maxNum}{WebGestaltR will exclude the categories with the number of annotated genes
larger than \code{maxNum} for enrichment analysis. The default is \code{500}.}

\item{fdrMethod}{For the ORA method, WebGestaltR supports five FDR methods: \code{holm},
\code{hochberg}, \code{hommel}, \code{bonferroni}, \code{BH} and \code{BY}. The default
is \code{BH}.}

\item{sigMethod}{Two methods of significance are available in WebGestaltR: \code{fdr} and
\code{top}. \code{fdr} means the enriched categories are identified based on the FDR and
\code{top} means all categories are ranked based on FDR and then select top categories
as the enriched categories. The default is \code{fdr}.}

\item{fdrThr}{The significant threshold for the \code{fdr} method. The default is \code{0.05}.}

\item{topThr}{The threshold for the \code{top} method. The default is \code{10}.}

\item{reportNum}{The number of enriched categories visualized in the final report. The default
is \code{20}. A larger \code{reportNum} may be slow to render in the report.}

\item{setCoverNum}{The number of expected gene sets after set cover to reduce redundancy.
It could get fewer sets if the coverage reaches 100\%. The default is \code{10}.}

\item{perNum}{The number of permutations for the GSEA method. The default is \code{1000}.}

\item{gseaP}{The exponential scaling factor of the phenotype score. The default is \code{1}.
When p=0, ES reduces to standard K-S statistics (See original paper for more details).}

\item{isOutput}{If \code{isOutput} is TRUE, WebGestaltR will create a folder named by
the \code{projectName} and save the results in the folder. Otherwise, WebGestaltR will
only return an R \code{data.frame} object containing the enrichment results. If
hundreds of gene list need to be analyzed simultaneously, it is better to set
\code{isOutput} to \code{FALSE}. The default is \code{TRUE}.}

\item{outputDirectory}{The output directory for the results.}

\item{projectName}{The name of the project. If \code{projectName} is \code{NULL},
WebGestaltR will use time stamp as the project name.}

\item{dagColor}{If \code{dagColor} is \code{binary}, the significant terms in the DAG
structure will be colored by steel blue for ORA method or steel blue (positive related)
and dark orange (negative related) for GSEA method. If \code{dagColor} is \code{continous},
the significant terms in the DAG structure will be colored by the color gradient based on
corresponding FDRs.}

\item{saveRawGseaResult}{Whether the raw result from GSEA is saved as a RDS file, which can be
 used for plotting. Defaults to \code{FALSE}. The list includes
 \describe{
   \item{Enrichment_Results}{A data frame of GSEA results with statistics}
   \item{Running_Sums}{A matrix of running sum of scores for each gene set}
   \item{Items_in_Set}{A list with ranks of genes for each gene set}
}}

\item{gseaPlotFormat}{The graphic format of GSEA enrichment plots. Either \code{svg},
\code{png}, or \code{c("png", "svg")} (default).}

\item{nThreads}{The number of cores to use for GSEA and set cover, and in batch function.}

\item{cache}{A directory to save data cache for reuse. Defaults to \code{NULL} and disabled.}

\item{hostName}{The server URL for accessing data. Mostly for development purposes.}

\item{useWeightedSetCover}{Use weighted set cover for ORA. Defaults to \code{TRUE}.}

\item{useAffinityPropagation}{Use affinity propagation for ORA. Defaults to \code{FALSE}.}

\item{usekMedoid}{Use k-medoid for ORA. Defaults to \code{TRUE}.}

\item{kMedoid_k}{The number of clusters for k-medoid. Defaults to \code{25}.}

\item{isMetaAnalysis}{whether to perform meta-analysis. Defaults to \code{TRUE}. FALSE is not currently implemented.}

\item{mergeMethod}{The method to merge the results from multiple omics (options: \code{mean}, \code{max}). Only used if \code{isMetaAnalysis = FALSE}. Defaults to \code{mean}.}

\item{normalizationMethod}{The method to normalize the results from multiple omics (options: \code{rank}, \code{median}, \code{mean}). Only used if \code{isMetaAnalysis = FALSE}.}

\item{referenceLists}{For the ORA method, users can also use an R object as the reference
gene list. \code{referenceLists} should be an R \code{vector} object containing the
reference gene list.}

\item{referenceListFiles}{For the ORA method, the users need to upload the reference gene
list. The extension of the \code{referenceListFile} should be \code{txt} and the file can
only contain one column: the reference gene list.}

\item{referenceTypes}{Vector of the ID types of the reference lists. The supported ID types
of WebGestaltR for the selected organism can be found by the function \code{listIdType}.
If the \code{organism} is \code{others}, users do not need to set this parameter.}

\item{referenceSets}{Users can directly select the reference sets from existing platforms in
WebGestaltR and do not need to provide the reference set through \code{referenceListFiles}.
All existing platforms supported in WebGestaltR can be found by the function
\code{listReferenceSets}. If \code{referenceListFiles} and \code{refereneceLists} are
\code{NULL}, WebGestaltR will use the \code{referenceSets} as the reference analyte sets.
Otherwise, WebGestaltR will use the user supplied reference set for enrichment analysis.
Must be a vector with length matching the input analyte list (i.e. c('genome', 'genome', 'KEGG'))}

\item{listNames}{The names of the analyte lists.}

\item{seed}{(optional) Integer seed for the GSEA permutations, making results
reproducible: the same input and the same seed give identical p-values and FDRs,
for the per-list results and the meta-analysis alike. Defaults to \code{NULL},
which seeds from the clock, so two runs of the same analysis will differ. Applies
to \code{enrichMethod = "GSEA"} only; the ORA path uses no randomness.}
}
\description{
Perform multi-omics analysis using WebGestaltR
}
