% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebGestaltRMultiomicsOra.R
\name{WebGestaltRMultiOmicsOra}
\alias{WebGestaltRMultiOmicsOra}
\title{Multi-omics ORA}
\usage{
WebGestaltRMultiOmicsOra(
  analyteLists = NULL,
  analyteListFiles = NULL,
  analyteTypes = NULL,
  enrichMethod = "ORA",
  organism = "hsapiens",
  enrichDatabase = NULL,
  enrichDatabaseFile = NULL,
  enrichDatabaseType = NULL,
  enrichDatabaseDescriptionFile = NULL,
  collapseMethod = "mean",
  minNum = 10,
  maxNum = 500,
  fdrMethod = "BH",
  sigMethod = "fdr",
  fdrThr = 0.05,
  topThr = 10,
  reportNum = 100,
  setCoverNum = 10,
  perNum = 1000,
  gseaP = 1,
  isOutput = TRUE,
  outputDirectory = getwd(),
  projectName = NULL,
  dagColor = "binary",
  nThreads = 1,
  cache = NULL,
  hostName = "https://www.webgestalt.org/",
  useWeightedSetCover = TRUE,
  useAffinityPropagation = FALSE,
  usekMedoid = FALSE,
  kMedoid_k = 25,
  referenceLists = NULL,
  referenceListFiles = NULL,
  referenceTypes = NULL,
  listNames = null
)
}
\description{
Multi-omics ORA
}
