% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMetaReport.R
\name{createMetaReport}
\alias{createMetaReport}
\title{createMetaReport}
\usage{
createMetaReport(
  hostName,
  outputDirectory,
  organism = "hsapiens",
  projectName,
  enrichMethod,
  geneSet_list,
  geneSetDes_list,
  geneSetDag_list,
  geneSetNet_list,
  interestingGeneMap_list,
  referenceGeneList_list,
  enrichedSig_list,
  geneTables_list,
  clusters_list,
  background_list,
  enrichDatabase_list = NULL,
  enrichDatabaseFile_list = NULL,
  enrichDatabaseType_list = NULL,
  enrichDatabaseDescriptionFile_list = NULL,
  interestGeneFile_list = NULL,
  interestGene_list = NULL,
  interestGeneType_list = NULL,
  collapseMethod = "mean",
  referenceGeneFile_list = NULL,
  referenceGene_list = NULL,
  referenceGeneType_list = NULL,
  referenceSet_list = NULL,
  minNum = 10,
  maxNum = 500,
  fdrMethod = "BH",
  sigMethod = "fdr",
  fdrThr = 0.05,
  topThr = 10,
  reportNum = 20,
  perNum = 1000,
  p = 1,
  dagColor = "binary",
  listNames = NULL
)
}
\description{
Generate HTML report for ORA and GSEA MetaAnalysis
}
\keyword{internal}
