% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatCheck.R
\name{formatCheck}
\alias{formatCheck}
\title{Check Format and Read Data}
\usage{
formatCheck(dataType = "list", inputGeneFile = NULL,
  inputGene = NULL)
}
\arguments{
\item{dataType}{Type of data, either \code{list}, \code{rnk} or \code{gmt}.
Could be \code{list}, \code{rnk} or \code{matrix} for \code{idToSymbol}.}

\item{inputGeneFile}{The data file to be mapped.}

\item{inputGene}{Or the input could be given as an R object.
GMT file should be read with \code{readGmt}.}
}
\value{
A list of data frame
}
\description{
Check Format and Read Data
}
