% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{gsea_rust}
\alias{gsea_rust}
\title{Run GSEA using rust library}
\usage{
gsea_rust(min_overlap, max_overlap, permutations, sets, parts, analytes, ranks)
}
\value{
List of the results of GSEA
}
\description{
Run GSEA using rust library
}
