% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listArchiveURL.R
\name{listArchiveUrl}
\alias{listArchiveUrl}
\alias{listArchiveURL}
\title{List WebGestalt Servers}
\usage{
listArchiveUrl()
}
\value{
A data frame of available servers.
}
\description{
List available WebGestalt servers.
}
