% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listOrganism.R
\name{listOrganism}
\alias{listOrganism}
\title{List Organisms}
\usage{
listOrganism(hostName = "http://www.webgestalt.org/")
}
\arguments{
\item{hostName}{The server URL for accessing data. Mostly for development purposes.}
}
\value{
A list of supported organisms.
}
\description{
List supported organisms on WebGestalt server.
}
