% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listOrganism.R
\name{listOrganism}
\alias{listOrganism}
\title{List Organisms}
\usage{
listOrganism(hostName = "http://www.webgestalt.org/", cache = NULL)
}
\arguments{
\item{hostName}{The server URL for accessing data. Mostly for development purposes.}

\item{cache}{A directory to save data cache for reuse. Defaults to \code{NULL} and disabled.}
}
\value{
A list of supported organisms.
}
\description{
List supported organisms on WebGestalt server.
}
