% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listReferenceSet.R
\name{listReferenceSet}
\alias{listReferenceSet}
\title{List Reference Sets}
\usage{
listReferenceSet(
  organism = "hsapiens",
  hostName = "https://www.webgestalt.org/",
  cache = NULL
)
}
\arguments{
\item{organism}{Currently, WebGestaltR supports 12 organisms. Users can use the function
\code{listOrganism} to check available organisms. Users can also input \code{others} to
perform the enrichment analysis for other organisms not supported by WebGestaltR. For
other organisms, users need to provide the functional categories, interesting list and
reference list (for ORA method). Because WebGestaltR does not perform the ID mapping for
the other organisms, the above data should have the same ID type.}

\item{hostName}{The server URL for accessing data. Mostly for development purposes.}

\item{cache}{A directory to save data cache for reuse. Defaults to \code{NULL} and disabled.}
}
\value{
A list of reference sets.
}
\description{
List available reference sets for the given organism on WebGestalt server.
}
