% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swGsea.R
\name{prepareGseaInput}
\alias{prepareGseaInput}
\title{Prepare input for standard GSEA}
\usage{
prepareGseaInput(rankFile, gmtFile)
}
\arguments{
\item{rankFile}{Path of the rnk file}

\item{gmtFile}{Path of the GMT file}
}
\value{
a data frame to be used in \code{swGsea}
}
\description{
A helper to read files for performing standard GSEA.
}
