% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGmt.R
\name{readGmt}
\alias{readGmt}
\title{Read GMT File}
\usage{
readGmt(gmtFile)
}
\arguments{
\item{gmtFile}{The file path or URL of the GMT file.}
}
\value{
A data frame with columns of "geneSet", "description", "gene".
}
\description{
Read GMT File
}
