\name{WebGestaltR-package}
\alias{WebGestaltR-package}

\docType{package}
\title{
	The R version of WebGestalt
}
\description{
	WebGestaltR is the R version of our well-known web application tool WebGestalt (www.webgestalt.org). The advantage of this R package is it can be integrated to other pipeline or simultaneous analyze multiple gene lists.
}
\details{
	\tabular{ll}{
		Package: \tab WebGestaltR\cr
		Type: \tab Package\cr
		License: \tab LGPL\cr
		LazyLoad: \tab yes\cr
	}
}

\author{
	Jing Wang
	Maintainer: Jing Wang <jing.wang@bcm.edu>
}

\references{
	WEB-based GEne SeT AnaLysis Toolkit (WebGestalt): update 2013. Nucleic Acids Res, 41 (Web Server issue), W77-83.
}

\seealso{
	\code{\link{WebGestaltR}}
}

\keyword{ package }
